refactoring out misc methods
