+++ prune_backups.go
@@ -5,38 +5,9 @@ import (
 	"fmt"
 	"os"
 	"path/filepath"
-	"runtime/debug"
 	"sort"
-	"strconv"
-	"strings"
 	"time"
 )
-
-var commitInfo = func() string {
-	//var version = "<unknown>"
-	var vcs_revision = "<unknown>"
-	var vcs_time = "<unknown>"
-	var vcs_modified = "<unknown>"
-	if info, ok := debug.ReadBuildInfo(); ok {
-		/*
-			if info.Main.Version != "" {
-				version = info.Main.Version // currently (Go 1.22.*) always returns "(devel)" - so ignore it. wait for https://github.com/golang/go/issues/50603 (ETA Go 1.24)
-			}
-		*/
-		for _, setting := range info.Settings {
-			if setting.Key == "vcs.revision" {
-				vcs_revision = setting.Value
-			}
-			if setting.Key == "vcs.time" {
-				vcs_time = setting.Value
-			}
-			if setting.Key == "vcs.modified" {
-				vcs_modified = setting.Value
-			}
-		}
-	}
-	return "rev " + vcs_revision + " from " + vcs_time + ", modified=" + vcs_modified
-}()
 
 func main() {
 
@@ -144,53 +115,6 @@ func pruneDirectory(pruneDirName string, now time.Time, toDeleteDirName string, 
 	}
 }
 
-func isFlagPassed(name string) bool {
-	found := false
-	flag.Visit(func(f *flag.Flag) {
-		if f.Name == name {
-			found = true
-		}
-	})
-	return found
-}
-
-func getAllButFirstMatchingPrefix(from []string, prefix string) []string {
-	var result = []string{} // make sure it's not nil
-	var first = true
-	for _, s := range from {
-		if strings.HasPrefix(s, prefix) {
-			if first {
-				first = false
-			} else {
-				result = append(result, s)
-			}
-		}
-	}
-	return result
-}
-
-func getAllMatchingPrefix(from []string, prefix string) []string {
-	var result = []string{} // make sure it's not nil
-	for _, s := range from {
-		if strings.HasPrefix(s, prefix) {
-			result = append(result, s)
-		}
-	}
-	return result
-}
-
-func getAllMatchingAllPrefixes(from []string, prefixes []string) []string {
-	var result = []string{} // make sure it's not nil
-	for _, s := range from {
-		for _, prefix := range prefixes {
-			if strings.HasPrefix(s, prefix) {
-				result = append(result, s)
-			}
-		}
-	}
-	return result
-}
-
 func createPrefixesForTimeSlotsToKeepOne(current time.Time) []string {
 	// Create an array to hold the prefixes
 	prefixes := make([]string, 24+30+119)
@@ -237,41 +161,6 @@ func createPrefixesForTimeSlotsToKeepOne(current time.Time) []string {
 	*/
 
 	return prefixes
-}
-
-func prevMonth(year *int, month *int) {
-	*month--
-	if *month <= 0 {
-		*month = 12
-		*year--
-	}
-}
-
-func toDateStr(year int, month int) string {
-	if month < 10 {
-		return strconv.Itoa(year) + "-0" + strconv.Itoa(month)
-	} else {
-		return strconv.Itoa(year) + "-" + strconv.Itoa(month)
-	}
-}
-
-func toDateStr3(year int, month int, day int) string {
-	strSep1 := "-"
-	strSep2 := "-"
-	if month < 10 {
-		strSep1 = "-0"
-	}
-	if day < 10 {
-		strSep2 = "-0"
-	}
-	return strconv.Itoa(year) + strSep1 + strconv.Itoa(month) + strSep2 + strconv.Itoa(day)
-}
-
-func twoDigit(i int) string {
-	if i < 10 {
-		return "0" + strconv.Itoa(i)
-	}
-	return strconv.Itoa(i)
 }
 
 func getFiltersForToday(current_time time.Time) []string {
@@ -332,14 +221,6 @@ func getFiltersFor30Dailys(current_time time.Time) []string {
 	return result
 }
 
-func daysInMonth(year int, month time.Month) int {
-	// Start with the first day of the next month
-	t := time.Date(year, month+1, 1, 0, 0, 0, 0, time.UTC)
-	// Subtract a day to get the last day of the original month
-	t = t.AddDate(0, 0, -1)
-	return t.Day()
-}
-
 func getMonthToLookForAnExtraMonthly(current_time time.Time) (bool, time.Time) {
 	day_of_today := current_time.Day()
 	month_before := get15thOfMonthBefore(current_time) // we don't really care for the exact day, just make sure it's not the 29th-31st as substracting a month will mean a hastle
@@ -398,12 +279,6 @@ func getMonthToLookForAnExtraMonthly(current_time time.Time) (bool, time.Time) {
 	// the 30 Dailys cover every day of the month before
 	is_needed := false
 	return is_needed, month_before
-}
-
-func get15thOfMonthBefore(current_time time.Time) time.Time {
-	t := time.Date(current_time.Year(), current_time.Month(), 15, 0, 0, 0, 0, time.UTC)
-	t = t.AddDate(0, -1, 0)
-	return t
 }
 
 func createPrefixesForTimeSlotsToKeepNone(current time.Time) []string {
